requirements.txt
flask 
numpy 
python-dotenv 
matplotlib 
fpdf 
requests
 groq
